requirements.txt
asgiref==3.3.0
Django==3.1.3
django-cleanup==5.1.0
Pillow==8.0.1
pymongo==3.11.0
pytz==2020.4
sqlparse==0.4.1
